Update the invitation SMS sending
This updates the InviteController to add support for the Swisscom REST SMS gateway and fix the phone number while resending the invitation. Three new environment variables are introduced: SMS_SWISSCOM_ACCOUNT for the Swisscom account ID, SMS_SWISSCOM_PASSWORD for the Swisscom account password and SMS_SWISSCOM_SENDER for the sender number to be displayed.
The SMS sending with Swisscom has been tested, but neither the SMS sending with OVH nor the email sending.